requirements.txt
pycountry==22.3.5
python-dotenv==0.21.1
requests==2.27.1
